Refactor code by removing unnecessary whitespace and comments; improve terminal mode handling in app selector and logging manager
