+++ src/TUI/core/components/__modal_win__.py
@@ -1,6 +1,5 @@
 import curses
 
-
 class ModalWindow:
     def __init__(self, stdscr):
         self.stdscr = stdscr
